+++ sphinx_simplepdf/__init__.py
@@ -4,9 +4,9 @@ from sphinx_simplepdf.directives.pdfinclude import PdfIncludeDirective
 
 
 def setup(app):
-    app.add_directive('if-builder', IfBuilderDirective)
-    app.add_directive('if-include', IfIncludeDirective)
-    app.add_directive('pdf-include', PdfIncludeDirective)
+    app.add_directive("if-builder", IfBuilderDirective)
+    app.add_directive("if-include", IfIncludeDirective)
+    app.add_directive("pdf-include", PdfIncludeDirective)
 
     return {
         "parallel_read_safe": True,