+++ sphinx_simplepdf/directives/ifinclude.py
@@ -1,11 +1,6 @@
-from docutils import nodes
-from docutils.parsers.rst import Directive
-
-
-from docutils import nodes
 from pathlib import Path
 
-
+from docutils.parsers.rst import Directive
 from sphinx.util import logging
 
 logger = logging.getLogger(__name__)
@@ -15,6 +10,7 @@ class IfIncludeDirective(Directive):
     """
     Directive to add a file based on builder.
     """
+
     has_content = True
     required_arguments = 1
     optional_arguments = 0
@@ -38,18 +34,15 @@ class IfIncludeDirective(Directive):
         builder = self.arguments[0]
 
         if self.env.app.builder.name.upper() == builder.upper():
-
             include_list = self.content
 
             # files get added in reversed order, probable cause is static target in doc, inserting element 1 at line
             # xy and then inserting element 2 at line xy leads to element 2 being in front of element 1
             # solution -> reverse list
             for file in reversed(include_list):
-
                 file = Path(file)
-                self.state_machine.insert_input([".. include:: " + str(file)],
-                                                self.state_machine.document.attributes["source"])
+                self.state_machine.insert_input(
+                    [".. include:: " + str(file)], self.state_machine.document.attributes["source"]
+                )
 
         return []
-
-