+++ sphinx_simplepdf/directives/pdfinclude.py
@@ -1,6 +1,5 @@
 from docutils import nodes
 from docutils.parsers.rst import Directive, directives
-
 from sphinx.util import logging
 
 logger = logging.getLogger(__name__)
@@ -10,6 +9,7 @@ class PdfIncludeDirective(Directive):
     """
     Directive to add content based on builder.
     """
+
     has_content = False
     required_arguments = 1
     optional_arguments = 0
@@ -22,7 +22,7 @@ class PdfIncludeDirective(Directive):
     }
 
     def __init__(self, *args, **kw):
-        self.pdf_specs = ''
+        self.pdf_specs = ""
 
         super().__init__(*args, **kw)
 
@@ -34,8 +34,8 @@ class PdfIncludeDirective(Directive):
         spec = self.options.get(option, default)
 
         if spec is not None:
-            if self.pdf_specs == '':
-                self.pdf_specs = '#'
+            if self.pdf_specs == "":
+                self.pdf_specs = "#"
 
             self.pdf_specs += f"{name}={spec}&"
 
@@ -46,10 +46,10 @@ class PdfIncludeDirective(Directive):
         height = self.options.get("height", "400px")
         width = self.options.get("width", "100%")
 
-        self._add_spec('page', 'page')
-        self._add_spec('toolbar', 'toolbar')
+        self._add_spec("page", "page")
+        self._add_spec("toolbar", "toolbar")
 
         html_code = f'<iframe src="{pdf_file}{self.pdf_specs}" style="height: {height}; width: {width}"></iframe>'
-        node = nodes.raw('', html_code, format='html')
+        node = nodes.raw("", html_code, format="html")
 
         return [node]