+++ sphinx_simplepdf/themes/simplepdf_theme/__init__.py
@@ -1,10 +1,10 @@
 """Sphinx ReadTheDocs theme.
 From https://github.com/ryan-roemer/sphinx-bootstrap-theme.
 """
+
 from os import path
 
-
-__version__ = '0.1.0'
+__version__ = "0.1.0"
 __version_full__ = __version__
 
 
@@ -16,7 +16,7 @@ def get_html_theme_path():
 
 # See http://www.sphinx-doc.org/en/stable/theming.html#distribute-your-theme-as-a-python-package
 def setup(app):
-    app.add_html_theme('simplepdf_theme', path.abspath(path.dirname(__file__)))
+    app.add_html_theme("simplepdf_theme", path.abspath(path.dirname(__file__)))
     # app.add_css_file('styles/main.css')
 
     return {