+++ sphinx_simplepdf/writers/simplepdf.py
@@ -9,16 +9,15 @@ logger = logging.getLogger(__name__)
 
 
 class SimplepdfTranslator(HTML5Translator):
-
     def get_secnumber(self, node: Element) -> tuple[int, ...] | None:
-        if node.get('secnumber'):
-            return node['secnumber']
+        if node.get("secnumber"):
+            return node["secnumber"]
 
         if isinstance(node.parent, nodes.section):
             docname = self.docnames[-1]
-            anchorname = "{}/#{}".format(docname, node.parent['ids'][0])
+            anchorname = "{}/#{}".format(docname, node.parent["ids"][0])
             if anchorname not in self.builder.secnumbers:
-                anchorname = "%s/" % docname  # try first heading which has no anchor
+                anchorname = f"{docname}/"  # try first heading which has no anchor
 
             if self.builder.secnumbers.get(anchorname):
                 return self.builder.secnumbers[anchorname]
@@ -29,36 +28,36 @@ class SimplepdfTranslator(HTML5Translator):
         secnumber = self.get_secnumber(node)
         if secnumber:
             self.body.append('<span class="section-number">')
-            self.body.append('.'.join(map(str, secnumber)) + self.secnumber_suffix)
-            self.body.append('</span>')
+            self.body.append(".".join(map(str, secnumber)) + self.secnumber_suffix)
+            self.body.append("</span>")
 
     def get_fignumber(self, node: Element) -> str | None:
-        figtype = self.builder.env.domains['std'].get_enumerable_node_type(node)
+        figtype = self.builder.env.domains["std"].get_enumerable_node_type(node)
         if not figtype:
             return None
 
-        if len(node['ids']) == 0:
-            msg = __('Any IDs not assigned for %s node') % node.tagname
+        if len(node["ids"]) == 0:
+            msg = __("Any IDs not assigned for %s node") % node.tagname
             logger.warning(msg, location=node)
             return None
 
         key = f"{self.docnames[-1]}/{figtype}"
-        figure_id = node['ids'][0]
+        figure_id = node["ids"][0]
         if figure_id not in self.builder.fignumbers.get(key, {}):
             return None
 
         prefix = self.config.numfig_format.get(figtype)
         if prefix is None:
-            msg = __('numfig_format is not defined for %s') % figtype
+            msg = __("numfig_format is not defined for %s") % figtype
             logger.warning(msg)
             return None
 
         numbers = self.builder.fignumbers[key][figure_id]
-        return prefix % '.'.join(map(str, numbers))
+        return prefix % ".".join(map(str, numbers))
 
     def add_fignumber(self, node: Element) -> None:
         fignumber = self.get_fignumber(node)
         if fignumber:
             self.body.append('<span class="caption-number">')
-            self.body.append(fignumber + ' ')
-            self.body.append('</span>')
+            self.body.append(fignumber + " ")
+            self.body.append("</span>")